% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{date_time_stamp}
\alias{date_time_stamp}
\title{Helper to generate a character with the current date and time}
\usage{
date_time_stamp()
}
\value{
s_chr: a scalar character with the current <date>_<time>
}
\description{
Useful for stamping a filename or log entry.
}
