% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_utils.R
\name{DF.assess_key}
\alias{DF.assess_key}
\title{Title}
\usage{
DF.assess_key(df, keyCol = 1L)
}
\arguments{
\item{df}{data.frame}

\item{keyCol}{a character or integer column index}
}
\value{
data.frame
}
\description{
Title
}
\examples{
seqs_df \%>\% DF.assess_key('sequence') \%>\% summary
}
