% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_utils.R
\name{DF.assess_missingness}
\alias{DF.assess_missingness}
\title{Vizualize and Quantify Data Frame Missingness Patterns}
\usage{
DF.assess_missingness(df, ..., title = "Patterns of Data Missingness")
}
\arguments{
\item{df}{data.frame with missing data}

\item{...}{additional arguments to heatmap, like \code{main} title}
}
\value{
NULL; called for side effect
}
\description{
Displays a cohort missingness heatmap and prints the % of missing data
 and complete cases
}
\examples{
DF.assess_missingness(mtcars)
}
