% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{gg_find_aes}
\alias{gg_find_aes}
\title{Query the available aesthetics for a ggplot geom}
\usage{
gg_find_aes(geom = "point")
}
\arguments{
\item{geom}{}
}
\value{
data frame
}
\description{
Query the available aesthetics for a ggplot geom
}
\examples{
gg_find_aes()
}
\seealso{
Other gg_aes: \code{\link{gg_geoms_aesthetics_mtx}}
}
