% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{gg_geoms_aesthetics_mtx}
\alias{gg_geoms_aesthetics_mtx}
\title{Build lLUT to associate ggplot2 layer with aesthetics.}
\usage{
gg_geoms_aesthetics_mtx()
}
\value{
data.frame
}
\description{
Build lLUT to associate ggplot2 layer with aesthetics.
}
\examples{
geoms_aesthetics_mtx()
}
\seealso{
Other gg_aes: \code{\link{gg_find_aes}}
}
