% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection_methods.R
\name{inspect_pkg}
\alias{inspect_pkg}
\title{Characterize and inspect contents of packages}
\usage{
inspect_pkg(pkg)
}
\arguments{
\item{pkg}{character(1)}
}
\value{
None, called for side effects.
}
\description{
Characterize and inspect contents of packages
}
\examples{
inspect_pkg("base")
}
