% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_utils.R
\name{mapnames}
\alias{mapnames}
\title{Map Names Flexibly in a Data.Frame or Vector
A wrapper for `mapvalues` (which only supports vectors, not data frames),
 with the same mapping rules.
`from` and `to` must have equal lengths.}
\usage{
mapnames(df, from, to)
}
\arguments{
\item{df}{: a data frame whose names need changed}

\item{from}{: a character vector with old col names.}

\item{to}{: a character vector with new col names.}
}
\value{
df : a dataframe with the same values as input, but possibly replaced colnames
}
\description{
Map Names Flexibly in a Data.Frame or Vector
A wrapper for `mapvalues` (which only supports vectors, not data frames),
 with the same mapping rules.
`from` and `to` must have equal lengths.
}
\examples{
# DF Operand
mapnames(mtcars,
         from = c("cyl", "hp"),
         to   = c("cylinders", "horsepower"))
# Nmd Vector Operand
mapnames(mtcars$cyl \%>\% set_names(),
         from = c(4, 6),
         to   = c("four", "six"))
}
