% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyUtils-package.R
\docType{package}
\name{MyUtils}
\alias{MyUtils}
\alias{MyUtils-package}
\alias{MyUtils-package}
\title{MyUtils.}
\description{
An generalized assortment of tools for dev workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.nak.co/mabadgeley/MyUtils}
  \item Report bugs at \url{https://git.nak.co/mabadgeley/MyUtils/issues}
}

}
