% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{stem}
\alias{stem}
\title{Extract the 'stem' from a file path (the filename without the extension).}
\usage{
stem(x_chr)
}
\arguments{
\item{x_chr}{character vector of file paths}
}
\value{
x_chr same length as x_chr of file path stems
}
\description{
Useful for providing unique names to filepath vectors
}
