% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partials.R
\name{str_}
\alias{str_}
\title{Wrapper for \code{\link[utils]{str}}}
\usage{
str_(object, list.len = 10, max.level = 2, strict.width = "cut", ...)
}
\arguments{
\item{object}{objects to inspect}

\item{...}{args passed to str}
}
\value{
side effect result of `str()` call
}
\description{
Wrapper for \code{\link[utils]{str}}
}
\examples{
s(mtcars)
s(mtcars, list.len=Inf)
}
