% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_case_camel}
\alias{str_case_camel}
\title{Convert string to camel case.}
\usage{
str_case_camel(x_chr, lower = TRUE)
}
\arguments{
\item{x_chr}{A character vector with snake_case_formatting}
}
\value{
A character vector the same len as x_chr, with CamelCaseFormatting
}
\description{
Transform a character vector into a camel case representation.
First, convert `x_chr` to title case with \code{\link{str_case_title}}
Second, remove spaces
Third, lower first character of each str if `lower`
}
\examples{
str_case_camel(x_chr = c("from_snake_case", "From Title Case"))
str_case_camel(x_chr = c("from_snake_case", "From Title Case"), lower=FALSE)
}
