% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str_subset}
\alias{str_subset}
\title{Wrapper for stringr to retain names.}
\usage{
str_subset(x_chr, pattern)
}
\arguments{
\item{x_chr}{a vector of characters to subset}

\item{pattern}{character to match by}
}
\value{
character with equal len or shorter than x_chr
}
\description{
mimicks the behavior of \code{\link[stringr]{str_subset}}
but works through \code{\link[stringr]{str_detect}}
}
\examples{
str_subset(letters \%>\% purrr::set_names(), rebus::or("a", "b"))
}
