% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyUtils-package.R
\docType{data}
\name{test_suite}
\alias{test_suite}
\title{Environment containing prototypical objects for testing.}
\format{An object of class \code{environment} of length 5.}
\usage{
test_suite
}
\description{
schema: <mode>_<type>
}
\keyword{datasets}
