% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_utils.R
\name{zero_dates}
\alias{zero_dates}
\title{Transform a vector of dates into a numeric vector of days since the first date.}
\usage{
zero_dates(x_date)
}
\arguments{
\item{x_date}{a vector of absolute dates to be transformed}
}
\value{
delta_date, a numeric vector of days since the first date
}
\description{
Helpful for masking absolute dates and getting a continuous predictor/target
}
\examples{
past_101_days <- Sys.Date() - seq(from = 0, to=100)
assert(all(zero_dates(past_101_days) == 100:0))
}
